Log all errors from cache.GetInformer call
Currently, `kind.Start` prints the last error from the polling process,
except for the NoKindMatchError. This behavior makes it users hard to
debug the cause of the `getInformer` call failure since it always
returns the context timeout error, which is the last error while
polling.

This commit logs all returned errors while calling `getInformer` in the
polling process.

Signed-off-by: Sunghoon Kang <hoon@linecorp.com>