:sparkles: Allow `builder.WebhookBuilder` to be used without a Manager
It can be beneficial to set up an external webhook server for a manager
or to even run the webhook server as separate deployment.
However, `builder.WebhookBuilder` is tied to a `manager.Manager`, which
you don't have when setting up webhook server of its own. With this
change, a server and a scheme can be supplied instead of a manager, so
the same functionality can be used without a manager while keeping
backwards compatibility.