client: preserve Raw field in CreateOptions, DeleteOptions, ListOptions and PatchOptions
The fields of the `Raw` metav1 Options are always been overwritten in the `AsCreateOptions`, `AsDeleteOptions`, `ListOptions` and  `PatchOptions`
functions even when these are not set in the client Options.

This PR fixes the implementation of this functions to preserve override the Raw fields only if the corresponding ones in the client Options are set.

Closes https://github.com/kubernetes-sigs/controller-runtime/issues/1984.