✨ Add indexes to fake client to mimic field selectors (#2025)
* Add indexes to fake client

Allow registration of indexes into the fake client to allow usage
of field selectors when doing a List. The indexing is done lazily
by doing a normal list and then filtering the results by computing
the index value for each list item.

To enable the main change for this commit, some refactorings are
performed: an internal and unexported function that checks
whether a field selector is in the form key=val or key==val is made
internal and exported to be shared between real and faked code to
ensure loyalty. Unit tests for it are added.

Co-authored-by: Paul Eichler <peichler@anynines.com>

* Address review comments

Co-authored-by: Paul Eichler <peichler@anynines.com>