client/cache/client_cache.go should be called resources
This file and bits in it were confusing, the client package doesn't have
any sort of built-in cache (as in, watch/list cache from informers).
Every request goes to the API server directly instead. The "internal
cache" is really just a storage for all rest client resources, this
commit changes that to clarify and avoid confusion.

Signed-off-by: Vince Prignano <vince@prigna.com>