Simplify cache.objectTypeForListObject
This changeset streamlines and simplifies the logic inside
objectTypeForListObject. This method is used to retrieve from a List
object, its respective actual resource GVK and a new instance of the
object.

In the previous iteration, the use of reflection was complicated to
follow and looking at the fields in the struct passed in, instead of
using the information we already have in the Scheme to create a new
instance of the expected object and return.

In addition to all of the above, we're also now handling
PartialObjectMetadataList(s).

Signed-off-by: Vince Prignano <vincepri@redhat.com>