:bug: builder: handle core Group correctly in webhook paths
Webhooks for the legacy/core v1 Kinds registered using the builder gives
paths like "/mutate--v1-pod" which may not be very intuitive to use in
the markers.

To follow the schema.GroupVersion().String() convention, just leave empty
Group out of the webhook path to get "/mutate-v1-pod" (also as it
was in the examples/builtin originally).

Signed-off-by: Mikko Ylinen <mikko.ylinen@intel.com>