:sparkling: Add terminal error
This change adds a terminal error into the reconcile package. The
purpose of that error is to still be logged and used in metrics but
avoid retrying in situations where it is known that that will not help.
It also adds a new metric for just terminal errors, as they often
indicate that some kind of human intervention is needed.