Use object's TypeMeta as GVK fallback
GVKForObject looks up the GVK for an Object via a Scheme, unless the
object is one of a few generic types. When handling multiple similar
resources, it's connivent to coerce those resource to a common static
type (often a subset of the actual object). These types cannot be
registered in the scheme as they do not map to a single GVK.

I would like to be able to resolve a GVK from a non-scheme registered
object type so that these types can be used in other useful methods that
internally call GVKForObject, like SetControllerReference. To do this,
if the lookup from the scheme fails, it will use the TypeMeta fields on
the object and return those values as the GVK. This behavior requires
the APIVersion and Kind fields are set, otherwise it will fall back to
returning the scheme error.

Signed-off-by: Scott Andrews <andrewssc@vmware.com>