:bug: Fix status subresource getting updated on Update when it is empty
Before this, Update and Patch requests will override to-level status
fields for objects that have a status resource if the status was
initially empty.

This happens because the code to set the status to what we already have
in the unupdated object converts the object into a `map[string]any`
using json serialization/deserialization, copies the original `status`
field, then converts back. The json deserializer unfortunatelly does not
properly clear the target, allowing data to remain that is absent in the
json serialization.