🐛 Handle unstructured status update with fake client (#2495)
* Handle unstructured status update with fake client

In order to prevent unintentional mutations outside of the status during
a status update, the non-status fields are copied back onto the passed
object.

This operation now gracefully handles both unstructured and typed
objects. Previously, it would panic if an unstructured object was passed
for a GVK known to the scheme, as internally the object within the
tracker is converted to the typed equivalent. The two types cannot
be directly assigned to each other and instead must be copied.

Signed-off-by: Scott Andrews <andrewssc@vmware.com>

* Review feedback

Using typed objects for the initial and actual object content assertion.
Unstructured objects are only used for the update.

Signed-off-by: Scott Andrews <andrewssc@vmware.com>

---------

Signed-off-by: Scott Andrews <andrewssc@vmware.com>