running: Make client.MatchingLabels faster
The 99% use-case of this is to set a selector, not to adjust an existing
one. This change introduces a fastpath that does that with half the
allocations and in a bit less than half the time.

The reason slowpath is slow is that for each label a requirement has to
be constructed that is then appended to a slice, both of which cause
allocations.