envtest: improve process cleanup
When envtest creates new processes, it wasn't setting the process group
to the parent, when the main process is killed or it panics, all the
other processes are orphaned, and need to be killed separately.

In addition, if for some reason we can't shutdown a process cleanly, we
should try to SIGKILL before returning from Stop.

Signed-off-by: Vince Prignano <vincepri@redhat.com>