:warning: Drop DiscoveryRESTMapper
The discovery restmapper was unused except for two places that we missed
(when creating a cache directly as opposed through the manager and in a
test). The discovery restmapper only ever loads mappings once during
startup and is unable to reload them if they change. It also isn't lazy
and we don't want anyone using it.