🌱 Remove the empty go.mod/sum files from hack/tools
All the tool binaries are managed via go install nowadays, so no need to
keep the module files around.

Add a .keep file to keep the now empty tools directory. It will still
be used by the build to store the tool binaries. Even though it would be
created automatically, having it as part of the git tree makes it more
obvious that it's still being used by the build.

See: 98472437 ("Use go-install for versioned dependencies")
Signed-off-by: Tom Wieczorek <twieczorek@mirantis.com>