🐛 prevent leader election when shutting down a non-elected manager
When leader election is enabled, a non-leader manager would never start the LeaderElection runnable group.
Thus, as the shutdown process calls the sync.Once Start func of the runnableGroup; it will start a new election.
This change ensures `Start` is ineffective during shutdown.

The test ensures the LeaderElection runnableGroup is not started during shutdown.

Signed-off-by: Alexandre Mahdhaoui <alexandre.mahdhaoui@gmail.com>