admission.Decoder is now an interface
The current implementation allows to create a decoder without
a scheme/codecs by just referencing the struct `&admission.Decoder{}`
given that codecs is private. The field was filled in before with the
inject package which has been removed. This change retains the Decoder
definition and makes it an interface, `admission.NewDecoder(scheme)` is
now the only way to instantiate our default decoder.

Signed-off-by: Vince Prignano <vincepri@redhat.com>