:bug: Cache: Keep selectors when byObject.Namespaces is defaulted
Prior to this patch, configuring for example a labelSelector in
`ByObject` and then inheriting namespaces from `DefaultNamespaces` meant
that the `labelSelector` would be ignored. This is because if namespaces
are configured, we set p a multinamespace cache. If we do that, we
expect each namespace entry to have the appropriate selectors
configured.

Unfortunately we defaulted the configs for`byObject.Namespaces` before
defaulting `byObject.Namespace` itself, causing the above-described
issue.

This change also adds a couple more tests for the cache defaulting.