:bug: Fix namespaced GVK check to use version (#2875)
* Fix namespaced GVK check to use version
A particular Kind may only be present in a specific version of a group. When
querying the RESTMapper we should include the version to ensure the
cached group is updated to pick up new versions as needed.

Signed-off-by: Griffin Davis <gcd@ibm.com>

* Add unit tests for IsGVKNamespaced

Include unit tests for varying combinations of new GVKs introduced at
runtime to validate cache updates.

Signed-off-by: Griffin Davis <gcd@ibm.com>

---------

Signed-off-by: Griffin Davis <gcd@ibm.com>