:sparkles: Workqueue: Add `controller` label
Currently, all controller-related metrics use the `controller` label to
identity the controller. The workqueue metrics however use the `name`
label instead. This is very confusing and unintuitive. Add the
`controller` label to the workqueue as well with the same value as the
existing `name` label. This keeps the cardinality the same, is
backwards-compatible and allows using one label throughout for
filtering.