:sparkles: Use aggregated discovery if available
This change makes the `RESTMapper` use [AggredatedDiscovery][0] if
available. `AggregatedDiscovery` is beta and thus enabled by default
since Kube 1.28.

What this means in particular is that during startup, we will always do
a request to `/api` and `/apis`. If `AggregatedDiscovery` is enabled,
that is all we have to do. If not, we have to do a third request.

The behavior for reloading remains the same: Do one request unless the
request didn't include a version. In that case we have to find the
group. If the group is unknown, we will keep on doing a request to `/api`
and one to `/apis` to find it. If it is found we are again done in the
`AggregatedDiscovery` case and have to do a third request otherwise.

All in all this means that with `AggregatedDiscovery` enabled, the only
case where this is worse is if the client interacts with only one
`GroupVersion`, as we end up doing one additional api request. If it is
disabled, we effectively waste two api requests.

[0]: https://github.com/kubernetes/enhancements/issues/3352