🌱 Do not deduplicate warnings by default
Controllers are long-running processes, and deduplication, as
implemented, increases memory use.

With this change, duplicate warnings will appear in the log by default.
However, this is safe, because Kubernetes rotates container logs by
default.

If a specific controller sees many duplicate warnings, it can configure
the handler to deduplicate them.