:bug: Error when source.Start() never returns
Contrary to everything else in controller-runtime, we expect
`source.Start` to be non-blocking. If someone implements a custom source
and gets this wrong, the resulting behavior is that the binary starts
successfully, but no reconciliation happens which is extremely difficult
to understand and debug.

This change makes us use the `CacheSyncTimeout` not only for the sources
`WaitForSync` but also for its `Start`.

It is worth noting that the current design of both requiring `Start` to
not block and `WaitForSync` to block is very confusing. It likely came
to be because we basicaly require two distinct contexsts in `Start`, one
to indicate the lifetime of the `Source` and one to indicate the `Start`
timeout.

To overall simplify and improve the code, the change also parallelizes
the `Start` of the sources.