:bug: Address review comments and fix metrics in priority queue
This change addresses Stefans review comments from the original PR and
fixes a bug in the metrics where we always included items that are not
ready yet - This is incorrect, metrics are only implemented on the
[basequeue][0] in upstream, meaning they are only emitted for items that
are ready. The impact of this was for example an incorrect queue_depth
metric.

[0]: https://github.com/kubernetes/kubernetes/blob/b1cb471982b74c13c26dbcc0f4e1b5ae92ea47e6/staging/src/k8s.io/client-go/util/workqueue/queue.go#L157