:warning: NewTypedUnmanaged: Stop requiring a manager
Currently, `NewUnmanaged` and `NewTypedUnmanaged` require a manager but
the only properties they ever access from it are `GetControllerOptions`
and `GetLogger`. It makes no sense to require a `Manager` for something
that is unmanaged, so remove that.

Doing so naively means that we require two options parameters for these
constructors, one that has the actual options and another one to default
from which again is confusing. Remove the options struct to default from
and instead implement a `DefaultFromConfig` that defaults the controller
options `config.Controller` which also clarifies what belongs where.
This defaulting gets automatically called in the managed constructors
and is exported so anyone can use it.

This change is breaking only for users of `New[Typed]Unmanaged` in that:
* The signature changes
* The defaulting of the `controller.Options` from `config.Controller`
  doesn't happen anymore, but it is still possible to do that explicitly