:bug: Fakeclient: Fix dataraces when writing to the scheme
We have a scheme write lock but plenty of other codeptaths that read
from the scheme and that don't do looking, resulting in dataraces if the
two happen in parallel.

This change introduces a simple RW lock and makes the fakeclient acquire
read locking for all its operations except when needing the write lock.
This isn't particularly smart, but given that we only have one codepath
that writes to the scheme, it seems good enough.