:warning: Fakeclient: Remove support for objects with pointer meta
This change removes supports for objects that have type or object meta
defined as pointer.

This is because the `FieldManagedObjectTracker` needed for support of
SSA in the fakeclient can not support it, as it has a lot of places
where it unconditionally accesses these fields using reflect, which will
cause panics.