:bug: Fakeclient: Validate managed fields on init objects
When the FieldManager in the FieldManagedObjectTracker encounters
managed fields it can not decode, it will just silently clear them.

This breaks anyone who has tests where the fakeclient gets initialized
with objects that have invalid managed fields. This worked prior to
using the `FieldManagedObjectTracker`, as the default
`testing.ObjectTracker` we used before has no understanding of them.

Validate the `managedFields` during fake client initialization to make
this issue properly visible.