Add sleep to manager test to avoid event broadcaster leak.
Event broadcaster goroutine leaks may happen due to a race condition
where the broadcaster is started and shutdown in quick successtion.

Realistically this is not an issue to usage since the broadcaster is
usually long lived.

To avoid this on the test, a short sleep has been added between start
and shutdown.

Signed-off-by: Borja Clemente <borja.clemente@gmail.com>