:seedling: Priorityqueue: Use separate b-trees for ready and non-ready items
This change refactors the priorityqueue to use distinct b-trees for
ready and non-ready items. This simplifies the code, because the two of
them effectively have different sorting requirements. We made this work
using one b-tree sorted by priority first and readyAt second and then
iterating through all priorities high to low until we find a ready item.
While correct, this is difficult to reason about.

The other issue is that currently, there is no explicit transition step
from non-ready to ready. As a result, we always had to iterate over all
items in case one got ready to then update metrics. With this change, we
instead have an explicit event when an item gets ready and we can then
update metrics for just this item and do not have to iterate over
everything.